requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
pydantic-settings>=2.0.0
passlib[bcrypt]
argon2-cffi
python-jose[cryptography]
email-validator
python-multipart
pandas
reportlab
openpyxl

